chore: Update Dockerfile to use official Node image
This commit updates the Dockerfile in the inkwell-frontend directory to use the official Node image. The previous line specifying the Node image has been removed and replaced with the same line. This change ensures that the correct Node image is used for building the inkwell-frontend application in the Docker container.