chore: Update package.json build script to use legacy OpenSSL provider
This commit modifies the build script in the package.json file of the inkwell-frontend directory. The "--openssl-legacy-provider" flag has been added to the "build" script for react-scripts. This change ensures that the legacy OpenSSL provider is used during the build process, which can help resolve any compatibility issues related to OpenSSL.